services/py-conductor/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.8
filelock==3.13.1
fsspec==2024.6.1
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.29.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==24.2
pillow==11.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvicorn-worker==0.3.0
watchfiles==1.0.4
websockets==14.2
